feat: Implement address autofill from zipcode. Includes Ruby/JS dependency updates.
