+++ Gemfile
@@ -60,7 +60,6 @@ gem 'tzinfo-data', platforms: [:mingw, :mswin, :x64_mingw, :jruby]
 
 gem 'devise'
 
-gem 'dotenv-rails'
 group :production do
   gem 'mysql2'
 end
