+++ app/views/layouts/application.html.erb
@@ -104,7 +104,6 @@
     <footer style="background-color: #f8f9fa; color: #6c757d;" class="text-center p-3 mt-auto">
       © PetTrack
     </footer>
-    <script src="https://cdn.jsdelivr.net/npm/jquery-jpostal-ja@2.14.45/jquery.jpostal.min.js"></script>
     <script async defer
       src="https://maps.googleapis.com/maps/api/js?key=<%= ENV['Maps_API_Key'] %>&callback=initMap&libraries=places,marker">
     </script>
