+++ package.json
@@ -11,6 +11,7 @@
     "bootstrap": "4.6.2",
     "dotenv": "^16.5.0",
     "jquery": "^3.7.1",
+    "jquery-jpostal-ja": "^2.14.69",
     "popper.js": "^1.16.1",
     "turbolinks": "^5.2.0",
     "webpack": "^4.46.0",
